requirements.txt
click>=6.7
python-digitalocean>=1.11
tox>=2.7.0
python-dateutil>=2.6.0
